requirements.txt
blinker==1.6.3
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.6
Flask-WTF==1.2.1
google-auth==2.23.3
google-auth-oauthlib==1.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==2.1.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
urllib3==2.0.6
Werkzeug==3.0.0
WTForms==3.0.1
